user details page wired-up
