--- src/utils/constants/demoData.ts
@@ -1,713 +1,0 @@
-import { IUserDetails } from "@/services/api/user";
-
-export const demoUser: IUserDetails = {
-    _id: '1',
-    user_name: 'john_doe',
-    name: 'John Doe',
-    email: 'johndoe@gmail.com',
-    mobile: '1234567890',
-    createdAt: '2021-10-01',
-    updatedAt: '2021-10-01',
-    leads: [],
-    recentlyViewedProducts: [
-        {
-            _id: "123",
-            image: "img",
-            brand_name: "aachho",
-            name: "product1",
-            price: 1400,
-            slug: "product-slug",
-        },
-        {
-            _id: "124",
-            image: "img",
-            brand_name: "aachho",
-            name: "product2",
-            price: 1400,
-            slug: "product-slug",
-        },
-        {
-            _id: "125",
-            image: "img",
-            brand_name: "aachho",
-            name: "product3",
-            price: 1400,
-            slug: "product-slug",
-        },
-        {
-            _id: "126",
-            image: "img",
-            brand_name: "aachho",
-            name: "product4",
-            price: 1400,
-            slug: "product-slug",
-        },
-    ],
-    cartItems: [
-        {
-            _id: "123",
-            image: "img",
-            brand_name: "aachho",
-            name: "product1",
-            price: 1400,
-            slug: "product-slug",
-            checkout_type: "buy_online",
-        },
-        {
-            _id: "124",
-            image: "img",
-            brand_name: "aachho",
-            name: "product2",
-            price: 1400,
-            slug: "product-slug",
-            checkout_type: "try_at_store",
-        },
-        {
-            _id: "125",
-            image: "img",
-            brand_name: "aachho",
-            name: "product3",
-            price: 1400,
-            slug: "product-slug",
-            checkout_type: "try_at_store",
-        },
-        {
-            _id: "126",
-            image: "img",
-            brand_name: "aachho",
-            name: "product4",
-            price: 1400,
-            slug: "product-slug",
-            checkout_type: "buy_online",
-        },
-        {
-            _id: "126",
-            image: "img",
-            brand_name: "aachho",
-            name: "product5",
-            price: 1400,
-            slug: "product-slug",
-            checkout_type: "buy_online",
-        },
-        {
-            _id: "126",
-            image: "img",
-            brand_name: "aachho",
-            name: "product6",
-            price: 1400,
-            slug: "product-slug",
-            checkout_type: "buy_online",
-        },
-        {
-            _id: "126",
-            image: "img",
-            brand_name: "aachho",
-            name: "product7",
-            price: 1400,
-            slug: "product-slug",
-            checkout_type: "buy_online",
-        },
-    ],
-    recentActivities: [
-        {
-            action_type: "user_login",
-            message: "Message",
-            createdAt: '2021-10-01'
-        },
-        {
-            action_type: "user_revisited_app",
-            message: "Message",
-            createdAt: '2021-10-01'
-        },
-        {
-            action_type: "user_login",
-            message: "Message",
-            createdAt: '2021-10-01'
-        }
-    ],
-    wishlist: [
-        {
-            _id: "123",
-            image: "img",
-            brand_name: "aachho",
-            name: "product1",
-            price: 1400,
-            slug: "product-slug",
-        },
-        {
-            _id: "124",
-            image: "img",
-            brand_name: "aachho",
-            name: "product2",
-            price: 1400,
-            slug: "product-slug",
-        },
-        {
-            _id: "125",
-            image: "img",
-            brand_name: "aachho",
-            name: "product3",
-            price: 1400,
-            slug: "product-slug",
-        },
-        {
-            _id: "126",
-            image: "img",
-            brand_name: "aachho",
-            name: "product4",
-            price: 1400,
-            slug: "product-slug",
-        },
-    ]
-}
-
-export const userBadges = [
-    {
-        title: "tag1",
-        createdAt: "2023-10-10",
-    },
-    {
-        title: "tag2",
-        createdAt: "2023-10-10",
-    },
-    {
-        title: "tag3",
-        createdAt: "2023-10-10",
-    },
-    {
-        title: "tag4",
-        createdAt: "2023-10-10",
-    },
-    {
-        title: "tag5",
-        createdAt: "2023-10-10",
-    },
-    {
-        title: "tag6",
-        createdAt: "2023-10-10",
-    },
-    {
-        title: "tag7",
-        createdAt: "2023-10-10",
-    },
-]
-
-export const demoUsers: IUserDetails[] = [
-    {
-        _id: '1',
-        user_name: 'john_doe',
-        name: 'John Doe',
-        email: 'johndoe@gmail.com',
-        mobile: '1234567890',
-        createdAt: '2021-10-01',
-        updatedAt: '2021-10-01',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [
-            {
-                _id: "123",
-                image: "img",
-                brand_name: "aachho",
-                name: "product1",
-                price: 1400,
-                slug: "product-slug",
-                checkout_type: "buy_online",
-            },
-            {
-                _id: "124",
-                image: "img",
-                brand_name: "aachho",
-                name: "product2",
-                price: 1400,
-                slug: "product-slug",
-                checkout_type: "try_at_store",
-            },
-        ],
-        recentActivities: [],
-        wishlist: [
-            {
-                _id: "123",
-                image: "img",
-                brand_name: "aachho",
-                name: "product1",
-                price: 1400,
-                slug: "product-slug",
-            },
-            {
-                _id: "124",
-                image: "img",
-                brand_name: "aachho",
-                name: "product2",
-                price: 1400,
-                slug: "product-slug",
-            },
-        ]
-    },
-    {
-        _id: '2',
-        user_name: 'jane_smith',
-        name: 'Jane Smith',
-        email: 'janesmith@gmail.com',
-        mobile: '9876543210',
-        createdAt: '2021-10-02',
-        updatedAt: '2021-10-02',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [
-            {
-                _id: "123",
-                image: "img",
-                brand_name: "aachho",
-                name: "product1",
-                price: 1400,
-                slug: "product-slug",
-                checkout_type: "buy_online",
-            },
-            {
-                _id: "124",
-                image: "img",
-                brand_name: "aachho",
-                name: "product2",
-                price: 1400,
-                slug: "product-slug",
-                checkout_type: "try_at_store",
-            },
-            {
-                _id: "123",
-                image: "img",
-                brand_name: "aachho",
-                name: "product1",
-                price: 1400,
-                slug: "product-slug",
-                checkout_type: "buy_online",
-            },
-            {
-                _id: "124",
-                image: "img",
-                brand_name: "aachho",
-                name: "product2",
-                price: 1400,
-                slug: "product-slug",
-                checkout_type: "try_at_store",
-            },
-        ],
-        recentActivities: [],
-        wishlist: [
-            {
-                _id: "123",
-                image: "img",
-                brand_name: "aachho",
-                name: "product1",
-                price: 1400,
-                slug: "product-slug",
-            },
-            {
-                _id: "124",
-                image: "img",
-                brand_name: "aachho",
-                name: "product2",
-                price: 1400,
-                slug: "product-slug",
-            },
-            {
-                _id: "125",
-                image: "img",
-                brand_name: "aachho",
-                name: "product3",
-                price: 1400,
-                slug: "product-slug",
-            },
-            {
-                _id: "126",
-                image: "img",
-                brand_name: "aachho",
-                name: "product4",
-                price: 1400,
-                slug: "product-slug",
-            },
-        ],
-    },
-    {
-        _id: '3',
-        user_name: 'alice_jones',
-        name: 'Alice Jones',
-        email: 'alicejones@gmail.com',
-        mobile: '5551112233',
-        createdAt: '2021-10-03',
-        updatedAt: '2021-10-03',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [
-            {
-                _id: "123",
-                image: "img",
-                brand_name: "aachho",
-                name: "product1",
-                price: 1400,
-                slug: "product-slug",
-                checkout_type: "buy_online",
-            },
-            {
-                _id: "124",
-                image: "img",
-                brand_name: "aachho",
-                name: "product2",
-                price: 1400,
-                slug: "product-slug",
-                checkout_type: "try_at_store",
-            },
-            {
-                _id: "123",
-                image: "img",
-                brand_name: "aachho",
-                name: "product1",
-                price: 1400,
-                slug: "product-slug",
-                checkout_type: "buy_online",
-            },
-        ],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '4',
-        user_name: 'bob_anderson',
-        name: 'Bob Anderson',
-        email: 'bobanderson@gmail.com',
-        mobile: '3334445555',
-        createdAt: '2021-10-04',
-        updatedAt: '2021-10-04',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '5',
-        user_name: 'emily_wilson',
-        name: 'Emily Wilson',
-        email: 'emilywilson@gmail.com',
-        mobile: '9998887777',
-        createdAt: '2021-10-05',
-        updatedAt: '2021-10-05',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '6',
-        user_name: 'michael_smith',
-        name: 'Michael Smith',
-        email: 'michaelsmith@gmail.com',
-        mobile: '7776665555',
-        createdAt: '2021-10-06',
-        updatedAt: '2021-10-06',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '7',
-        user_name: 'sara_brown',
-        name: 'Sara Brown',
-        email: 'sarabrown@gmail.com',
-        mobile: '2223334444',
-        createdAt: '2021-10-07',
-        updatedAt: '2021-10-07',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '8',
-        user_name: 'adam_taylor',
-        name: 'Adam Taylor',
-        email: 'adamtaylor@gmail.com',
-        mobile: '8889990000',
-        createdAt: '2021-10-08',
-        updatedAt: '2021-10-08',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '9',
-        user_name: 'lisa_miller',
-        name: 'Lisa Miller',
-        email: 'lisamiller@gmail.com',
-        mobile: '4445556666',
-        createdAt: '2021-10-09',
-        updatedAt: '2021-10-09',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '10',
-        user_name: 'peter_jackson',
-        name: 'Peter Jackson',
-        email: 'peterjackson@gmail.com',
-        mobile: '1112223333',
-        createdAt: '2021-10-10',
-        updatedAt: '2021-10-10',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '11',
-        user_name: 'olivia_davis',
-        name: 'Olivia Davis',
-        email: 'oliviadavis@gmail.com',
-        mobile: '6667778888',
-        createdAt: '2021-10-11',
-        updatedAt: '2021-10-11',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '12',
-        user_name: 'david_clark',
-        name: 'David Clark',
-        email: 'davidclark@gmail.com',
-        mobile: '2224446666',
-        createdAt: '2021-10-12',
-        updatedAt: '2021-10-12',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '13',
-        user_name: 'amy_carter',
-        name: 'Amy Carter',
-        email: 'amycarter@gmail.com',
-        mobile: '7778889999',
-        createdAt: '2021-10-13',
-        updatedAt: '2021-10-13',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '14',
-        user_name: 'ryan_hall',
-        name: 'Ryan Hall',
-        email: 'ryanhall@gmail.com',
-        mobile: '1115559999',
-        createdAt: '2021-10-14',
-        updatedAt: '2021-10-14',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '15',
-        user_name: 'natalie_ward',
-        name: 'Natalie Ward',
-        email: 'natalieward@gmail.com',
-        mobile: '4446668888',
-        createdAt: '2021-10-15',
-        updatedAt: '2021-10-15',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '16',
-        user_name: 'jason_martin',
-        name: 'Jason Martin',
-        email: 'jasonmartin@gmail.com',
-        mobile: '9992223333',
-        createdAt: '2021-10-16',
-        updatedAt: '2021-10-16',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '17',
-        user_name: 'elizabeth_baker',
-        name: 'Elizabeth Baker',
-        email: 'elizabethbaker@gmail.com',
-        mobile: '3336669999',
-        createdAt: '2021-10-17',
-        updatedAt: '2021-10-17',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '18',
-        user_name: 'mark_morgan',
-        name: 'Mark Morgan',
-        email: 'markmorgan@gmail.com',
-        mobile: '8881112222',
-        createdAt: '2021-10-18',
-        updatedAt: '2021-10-18',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '19',
-        user_name: 'victoria_hill',
-        name: 'Victoria Hill',
-        email: 'victoriahill@gmail.com',
-        mobile: '5557779999',
-        createdAt: '2021-10-19',
-        updatedAt: '2021-10-19',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '20',
-        user_name: 'sam_jackson',
-        name: 'Sam Jackson',
-        email: 'samjackson@gmail.com',
-        mobile: '5551234567',
-        createdAt: '2021-10-20',
-        updatedAt: '2021-10-20',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '21',
-        user_name: 'grace_kelly',
-        name: 'Grace Kelly',
-        email: 'gracekelly@gmail.com',
-        mobile: '1112223333',
-        createdAt: '2021-10-21',
-        updatedAt: '2021-10-21',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '22',
-        user_name: 'daniel_brown',
-        name: 'Daniel Brown',
-        email: 'danielbrown@gmail.com',
-        mobile: '9998887777',
-        createdAt: '2021-10-22',
-        updatedAt: '2021-10-22',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '23',
-        user_name: 'oliver_morris',
-        name: 'Oliver Morris',
-        email: 'olivermorris@gmail.com',
-        mobile: '5556667777',
-        createdAt: '2021-10-23',
-        updatedAt: '2021-10-23',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '24',
-        user_name: 'megan_turner',
-        name: 'Megan Turner',
-        email: 'meganturner@gmail.com',
-        mobile: '3334445555',
-        createdAt: '2021-10-24',
-        updatedAt: '2021-10-24',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [],
-    },
-    {
-        _id: '25',
-        user_name: 'charlie_wilson',
-        name: 'Charlie Wilson',
-        email: 'charliewilson@gmail.com',
-        mobile: '7778889999',
-        createdAt: '2021-10-25',
-        updatedAt: '2021-10-25',
-        leads: [],
-        recentlyViewedProducts: [],
-        cartItems: [],
-        recentActivities: [],
-        wishlist: [
-            {
-                _id: "123",
-                image: "img",
-                brand_name: "aachho",
-                name: "product1",
-                price: 1400,
-                slug: "product-slug",
-            },
-            {
-                _id: "124",
-                image: "img",
-                brand_name: "aachho",
-                name: "product2",
-                price: 1400,
-                slug: "product-slug",
-            },
-            {
-                _id: "125",
-                image: "img",
-                brand_name: "aachho",
-                name: "product3",
-                price: 1400,
-                slug: "product-slug",
-            },
-            {
-                _id: "126",
-                image: "img",
-                brand_name: "aachho",
-                name: "product4",
-                price: 1400,
-                slug: "product-slug",
-            },
-        ]
-    },
-];
